+++ CHANGELOG.md
@@ -7,6 +7,10 @@ A [BC BREAK] means the update will break the project for many reasons:
 * new dependencies
 * class refactoring
 
+### 2014-03-12
+
+* [BC BREAK] SecurityPolicyContenairAware has been renamed to SecurityPolicyContainerAware (typo).
+
 ### 2013-07-26
 
 * remove the sonata_formatter_type_selector and it is now named sonata_formatter_type